+++ r2-streamer-swift/Streamer.swift
@@ -16,11 +16,14 @@ import R2Shared
 public final class Streamer: Loggable {
     
     /// Creates the default parsers provided by Readium.
-    public static func makeDefaultParsers(pdfFactory: PDFDocumentFactory) -> [PublicationParser] {
+    public static func makeDefaultParsers(
+        pdfFactory: PDFDocumentFactory = DefaultPDFDocumentFactory(),
+        httpClient: HTTPClient = DefaultHTTPClient()
+    ) -> [PublicationParser] {
         [
             EPUBParser(),
             PDFParser(pdfFactory: pdfFactory),
-            ReadiumWebPubParser(pdfFactory: pdfFactory),
+            ReadiumWebPubParser(pdfFactory: pdfFactory, httpClient: httpClient),
             ImageParser(),
             AudioParser()
         ]
@@ -47,19 +50,18 @@ public final class Streamer: Loggable {
         contentProtections: [ContentProtection] = [],
         archiveFactory: ArchiveFactory = DefaultArchiveFactory(),
         pdfFactory: PDFDocumentFactory = DefaultPDFDocumentFactory(),
+        httpClient: HTTPClient = DefaultHTTPClient(),
         onCreatePublication: Publication.Builder.Transform? = nil
     ) {
-        self.parsers = parsers + (ignoreDefaultParsers ? [] : Streamer.makeDefaultParsers(pdfFactory: pdfFactory))
+        self.parsers = parsers + (ignoreDefaultParsers ? [] : Streamer.makeDefaultParsers(pdfFactory: pdfFactory, httpClient: httpClient))
         self.contentProtections = contentProtections
         self.archiveFactory = archiveFactory
-        self.pdfFactory = pdfFactory
         self.onCreatePublication = onCreatePublication
     }
     
     private let parsers: [PublicationParser]
     private let contentProtections: [ContentProtection]
     private let archiveFactory: ArchiveFactory
-    private let pdfFactory: PDFDocumentFactory
     private let onCreatePublication: Publication.Builder.Transform?
 
     /// Parses a `Publication` from the given asset.