+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+* `Streamer` takes a new optional `HTTPClient` dependency to handle HTTP requests.
+
+
 ### Changed
 
 * The HTTP server now requests that publication resources are not cached by browsers.